refactor(app): extract crypto group account row
