+++ packages/app/src/account/component/crypto-currency-group-accounts/crypto-currency-group-accounts.tsx
@@ -1,11 +1,8 @@
-import { AccountTypeEnum } from '@budgie/contracts';
 import { View } from 'react-native';
-
-import { isDefined } from '@rnw-community/shared';
 
 import { AccountRowInterface } from '../../interface/account-row.interface';
 import { CryptoCurrencyGroupInterface } from '../../interface/crypto-currency-group.interface';
-import { AccountGridItem } from '../account-grid-item/account-grid-item';
+import { CryptoCurrencyGroupAccountRow } from '../crypto-currency-group-account-row/crypto-currency-group-account-row';
 
 interface Props {
     readonly group: CryptoCurrencyGroupInterface;
@@ -24,23 +21,14 @@ const pairAccountsIntoRows = (group: CryptoCurrencyGroupInterface): AccountRowIn
     return rows;
 };
 
-const renderAccountRow = (row: AccountRowInterface) => {
-    const rightAccount = isDefined(row.right) ? (
-        <AccountGridItem account={row.right} type={AccountTypeEnum.CRYPTO} isLeft={false} />
-    ) : (
-        <View className="flex-1" />
-    );
-
-    return (
-        <View key={row.left.id} className="flex-row">
-            <AccountGridItem account={row.left} type={AccountTypeEnum.CRYPTO} isLeft />
-            {rightAccount}
-        </View>
-    );
-};
-
 export const CryptoCurrencyGroupAccounts = ({ group }: Props) => {
     const rows = pairAccountsIntoRows(group);
 
-    return <View className="gap-y-3">{rows.map(renderAccountRow)}</View>;
+    return (
+        <View className="gap-y-3">
+            {rows.map(row => (
+                <CryptoCurrencyGroupAccountRow key={row.left.id} row={row} />
+            ))}
+        </View>
+    );
 };
